commit2

+++ ft_printf.h
@@ -10,7 +10,7 @@ int ft_putchar(int c);
 int ft_putstr(char *str);
 int ft_putnbr(int n);
 int ft_putunbr(unsigned int n);
-int ft_puthex(unsigned int n, int upper);
+int ft_puthex(unsigned long long n , int upper);
 int ft_putptr(unsigned long n);
